DPUC/fetcher_search/requirements.txt
Django>=3.0,<4.0
djangorestframework
markdown
django-filter
mysql-connector-python
elasticsearch